MDBOOK: added glossary terms
Added the terms from https://developer.holochain.net/Glossary. I wrote an XSL script to convert it from XML (which is used in the Moodle devcamp) to markdown. I used an html definition list in the markdown, but let me know if it would be better to use a different pattern and I'll update the xsl and re-export.