Rewrite `&PathBuf`s in signatures to `&Path`s
Taking &PathBufs is like taking &Strings, if you can't extend it because
it's not mutable, there's no use making sure it's on the heap. Taking
&Paths or &strs is simply more flexible to the caller without losing
anything for the procedure.

In the standard library, all code that wants to take an &Path takes
an impl AsRef<Path> instead, this way Strings and &strs can be passed
ergonomically. This commit uses &Path generally, because it may not be
worth the increase in binary size to add a generic. The author has no
preference one way or another and is open to rewriting to use the
alternative approach.