Resolve failing test by removing StateWrapper
The state is held by an Arc.
Fundamentally, this means that it will be dropped as soon as
no thread holds a reference to it.

The previous implementation of Drop for Instance assumed that
there were no other instances referencing the data it dropped.
Since we used a derived implementation of Clone and
the data was held in an Arc, it was possible to exploit by calling
`.clone()`, making a "shallow clone", and then drop one of the
instances, leaving the other one invalid and causing a panic if
it was used.