! DON'T SERIALIZE WIRE MESSAGE TWICE FOR SIGNATURES !
As we just found out, the serialization of WireMessages is not deterministic because of a hidden HashMap. The code fixed here assumed `message.into::<String/Opaque>()` to always return the same result since it's doing it once for the payload sent, and once for the calculation of the signature. The result would be (and almost certainly was) wrong signatures which make Sim2h server ignore the message.